use poll() to reduce CPU hammering
setup a poll on the mousefd; it will return every five seconds or on
input. once the return has happened, check for number of events
(mousepoll), and if there is an event (and input to be read), issue a
read() only when there is data. this should reduce CPU overhead.

will port this change to joystick shortly.